move post rendering logic from homepage to postview
